requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
dotenv==0.9.9
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
